//Updated Enemy bounce
The enemy bounce script uses the linear drag (rb.drag) to slow the object down.

If needed, a timer can be used to do this. I commented out the timer code for the sake of testing. DO NOT TOUCH IT!!!!

There is a public float so you can test out the rb.drag. in my opinion, 1 and 1.5 work the best