//PauseMenu is Done
When you hit enter, the pause menu appears! You can click resume or press enter again to resume and restart the level with the restart button. The quit button doesn't work in the editor, but it works in a build (that's not a mistake, that something done on purpose in Unity.

NOTE: This MIGHT replace the Press R to restart